% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugs.R
\name{getGeneDrugTargets}
\alias{getGeneDrugTargets}
\title{Get gene drug targets}
\usage{
getGeneDrugTargets(geneData, number = NULL)
}
\arguments{
\item{geneData}{A data frame of gene data that must column Hugo_Symbol.}

\item{number}{An integer specifying the number of most druggable genes to return.
Default is NULL which gets all genes.}
}
\value{
A list with two elements:
\itemize{
  \item mostDruggableGenes: A named vector of the most druggable genes and
   their frequency in the drug database.
  \item drugsThatTargetTheseGenes: A data frame of the drugs that target
   the most druggable genes with columns Gene, interaction_claim_source,
  interaction_types, drug_claim_name, drug_claim_primary_name, drug_name,
   drug_chembl_id, PMIDs
}
}
\description{
This function takes a data frame of gene data and returns a list of
the most druggable genes and information on the drugs that target them.
Gene druggability is found from the Drug Gene Interaction Database
 referenced below
}
\examples{
# Example 1:
# Use gene data provided from the maftools package
laml.maf <- system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf)
geneSummary <- maftools::getGeneSummary(laml)

getGeneDrugTargets(geneSummary)

}
\references{
Mayakonda A, Lin DC, Assenov Y, Plass C, Koeffler HP. 2018.
 Maftools: efficient and comprehensive analysis of somatic variants in cancer.
  Genome
Research. http://dx.doi.org/10.1101/gr.239244.118

Griffith, M., Griffith, O., Coffman, A. et al.
DGIdb: mining the druggable genome. Nat Methods 10, 1209–1210 (2013).
 https://doi.org/10.1038/nmeth.2689
}
