% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_drug_targets.R
\name{plotDruggableGenes}
\alias{plotDruggableGenes}
\title{Plot druggable genes}
\usage{
plotDruggableGenes(drugFrequencies)
}
\arguments{
\item{drugFrequencies}{A named vector of drug frequencies for each gene.}
}
\value{
A ggplot object of a bar plot of the number of drugs for each gene.
}
\description{
This function takes a named vector of drug frequencies for each gene and
 returns a bar plot of the number of drugs for each gene.
}
\examples{
# Example 1:
# Use gene data provided from the maftools package
laml.maf <- system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf)
geneSummary <- maftools::getGeneSummary(laml)

drugTargets <- DrugSeekR::getGeneDrugTargets(geneSummary)
plotDruggableGenes(drugTargets$mostDruggableGenes)

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
 Springer-Verlag New York, 2016.
}
