% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugSeekR.R
\name{runDrugSeekR}
\alias{runDrugSeekR}
\title{Launch Shiny App for DrugSeekR}
\usage{
runDrugSeekR()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for DrugSeekR.
The code for the app can be found in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

DrugSeekR::runDrugSeekR()
}

}
