% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationStatistics.R
\name{calculateMutationFrequencies}
\alias{calculateMutationFrequencies}
\title{Calculates Mutation Frequency and Percentage}
\usage{
calculateMutationFrequencies(geneData)
}
\arguments{
\item{geneData}{A list of genetic data returned from the maftools
getGeneSummary function}
}
\value{
returns a list of two elements: the number of mutations for
   each gene and the mutation percentage for each gene.
}
\description{
A function that calculates
the frequency of observed mutations within a sample and their percentages
}
\examples{
# Example 1:
# Using maftools maf data available with package
laml.maf <- system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf)
geneSummary <- maftools::getGeneSummary(laml)

# Calculate Mutation Frequency and Percentage
mutationStatistics <- calculateMutationFrequencies(geneSummary)
mutationStatistics

}
\references{
Mayakonda A, Lin DC, Assenov Y, Plass C, Koeffler HP. 2018. Maftools:
efficient and comprehensive analysis of somatic variants in cancer. Genome
Research. http://dx.doi.org/10.1101/gr.239244.118
}
