% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_drug_targets.R
\name{plotDruggableGenes}
\alias{plotDruggableGenes}
\title{Plot druggable genes}
\usage{
plotDruggableGenes(drugFrequencies, number = NULL)
}
\arguments{
\item{drugFrequencies}{A named vector of drug frequencies for each gene.}

\item{number}{An integer specifying the number of genes to plot. Default is NULL which gets all genes.}
}
\value{
A ggplot object of a bar plot of the number of drugs for each gene.
}
\description{
This function takes a named vector of drug frequencies for each gene and
 returns a bar plot of the number of drugs for each gene.
}
\examples{
# Example 1:
# Use gene data provided from the maftools package. It contains a maf file
# with data from a paper studying adult acute myeloid leukemia
laml.maf <- system.file('extdata', 'tcga_laml.maf.gz', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf)
geneSummary <- maftools::getGeneSummary(laml)

drugTargets <- DrugSeekR::getGeneDrugTargets(geneSummary)
plotDruggableGenes(drugTargets$mostDruggableGenes, number = 5)

}
\references{
Cancer Genome Atlas Research, N. Genomic and epigenomic landscapes of adult de novo acute myeloid leukemia.
 N Engl J Med 368, 2059-74, 2013.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
 Springer-Verlag New York, 2016.
}
